- Spawn and Add Quest cheats
When you spend hours making your own decent sorta Item Database system and complete it, then find out about Unreal's built-in version that's better so you spend hours learning that one and using that instead...

But good job I learned this one coz know I know how to stream assets in and out. Can be useful to use less RAM in the future.

- AddQuest [QuestID] cheat. The IDs can be found in the Quest Data Assets.

- Spawn [ItemID]. Currently only Camera and Frog is registered. Will add more later.
(If you want Actors to be spawnable, it needs to be in its own BP and inherit from PrimaryAssetActor.